[src] Split modules of HiFi-GAN
